+++ testdata/panic.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"gitlab.com/tozd/go/errors"
+)
+
+func main() {
+	panic(errors.WithDetails(errors.Base("panic error"), "key", "value"))
+}